% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyCohortPersistenceCriteria.R
\name{applyCohortPersistenceCriteria}
\alias{applyCohortPersistenceCriteria}
\title{Apply persistence criteria.}
\usage{
applyCohortPersistenceCriteria(
  connectionDetails = NULL,
  connection = NULL,
  sourceCohortDatabaseSchema = NULL,
  sourceCohortTable,
  targetCohortDatabaseSchema = NULL,
  targetCohortTable,
  cdmDatabaseSchema,
  oldCohortId,
  newCohortId,
  tillEndOfObservationPeriod = FALSE,
  offsetCohortStartDate = NULL,
  offsetCohortEndDate = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  purgeConflicts = FALSE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{sourceCohortDatabaseSchema}{Schema name where your source cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{sourceCohortTable}{The name of the source cohort table.}

\item{targetCohortDatabaseSchema}{Schema name where your target cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{targetCohortTable}{The name of the target cohort table.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oldCohortId}{The cohort id of the cohort that needs to be modified.}

\item{newCohortId}{The cohort id of the output cohort.}

\item{tillEndOfObservationPeriod}{The cohort will persist till end of the overlapping observation period. An era logic
will be applied.}

\item{offsetCohortStartDate}{}

\item{offsetCohortEndDate}{Apply a fixed persistence criteria relative to cohort end date. A new cohort
end date will be created by adding persistence days to cohort_end_date with a
value that is minimum of the cohort_end_date + offsetCohortEndDate
or observation_period_end_date of the overlapping observation period. An era logic
will be applied.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{purgeConflicts}{If there are conflicts in the target cohort table i.e. the target cohort table
already has records with newCohortId, do you want to purge and replace them
with transformed. By default - it will not be replaced, and an error message is thrown.}
}
\description{
Apply cohort persistence criteria. Only one persistence criteria may be used at a time. The three options are a) persist till
end of observation period, b) persist for a certain number of fixed days after cohort_start_date, c) persist for
a certain number of fixed days after cohort_end_days. In all cases, the given cohort (oldCohortId) is treated as
an event and the criteria is applied to get new event dates. Event dates are converted to cohort dates by cohort
era fy routine in final step.

Offset: The event end date is derived from adding a number of days to the event's start or end date. If an offset is added to the
event's start date, all cohort episodes will have the same fixed duration (limited by duration of continuos observation).
If an offset is added to the event's end date, persons in the cohort may have varying cohort duration
times due to the varying event duration. This event persistence assures that the cohort end date will be no greater than
the selected index event date, plus the days offset.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
CohortAlgebra::applyCohortPersistenceCriteria(
  connection = connection,
  sourceCohortTable = "cohort",
  targetCohortTable = "cohort",
  oldCohortId = 3,
  newCohortId = 2,
  tillEndOfObservationPeriod = TRUE,
  purgeConflicts = TRUE
)
}

}
