% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyDemographicCriteria.R
\name{applyDemographicCriteria}
\alias{applyDemographicCriteria}
\title{Apply Demographic cohort}
\usage{
applyDemographicCriteria(
  connectionDetails = NULL,
  connection = NULL,
  sourceCohortDatabaseSchema = NULL,
  sourceCohortTable,
  targetCohortDatabaseSchema = NULL,
  targetCohortTable,
  cdmDatabaseSchema,
  oldCohortId,
  newCohortId,
  filterGenderConceptId = NULL,
  filterByAgeRange = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  purgeConflicts = TRUE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{sourceCohortDatabaseSchema}{Schema name where your source cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{sourceCohortTable}{The name of the source cohort table.}

\item{targetCohortDatabaseSchema}{Schema name where your target cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{targetCohortTable}{The name of the target cohort table.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oldCohortId}{The cohort id of the cohort that needs to be modified.}

\item{newCohortId}{The cohort id of the output cohort.}

\item{filterGenderConceptId}{Provide an array of integers corresponding to conceptId to look for in the gender_concept_id
field of the person table.}

\item{filterByAgeRange}{Provide an array of two values, where second value is >= first value to filter the persons age on cohort_start_date.
Age is calculated as YEAR(cohort_start_date) - person.year_of_birth}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{purgeConflicts}{If there are conflicts in the target cohort table i.e. the target cohort table
already has records with newCohortId, do you want to purge and replace them
with transformed. By default - it will not be replaced, and an error message is thrown.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
CohortAlgebra::applyDemographicCriteria(
  connection = connection,
  sourceCohortTable = tableName,
  oldCohortId = 3,
  newCohortId = 2,
  filterGenderConceptId = c(8201),
  purgeConflicts = TRUE
)
}

}
