% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CensorCohortDates.R
\name{censorCohortDates}
\alias{censorCohortDates}
\title{Censor cohort date}
\usage{
censorCohortDates(
  connectionDetails = NULL,
  connection = NULL,
  sourceCohortDatabaseSchema = NULL,
  sourceCohortTable,
  targetCohortDatabaseSchema = NULL,
  targetCohortTable,
  oldCohortId,
  newCohortId,
  cohortStartDateLeftCensor = NULL,
  cohortEndDateRightCensor = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  purgeConflicts = FALSE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{oldCohortId}{The cohort id of the cohort that needs to be modified.}

\item{newCohortId}{The cohort id of the output cohort.}

\item{cohortStartDateLeftCensor}{the minimum date for the cohort start.}

\item{cohortEndDateRightCensor}{the maximum date for the cohort end.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{purgeConflicts}{If there are conflicts in the target cohort table i.e. the target cohort table
already has records with newCohortId, do you want to purge and replace them
with transformed. By default - it will not be replaced, and an error message is thrown.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortDatabaseSchema}{Schema name where your cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}
}
\description{
Censor cohort date by right, left, both censoring. All cohorts will be truncated to the given date.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
CohortAlgebra::censorCohortDates(
  connection = connection,
  sourceCohortTable = 'cohort',
  targetCohortTable = 'cohort',
  oldCohortId = 3,
  newCohortId = 2,
  cohortStartDateLeftCensor = as.Date("2010-01-09"),
  purgeConflicts = TRUE
)
}

}
