% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortAlgebra-package.R, R/CohortAlgebra.R
\docType{package}
\name{CohortAlgebra-package}
\alias{CohortAlgebra-package}
\title{CohortAlgebra: Use of Interval Algebra to Create New Cohort(s) from Existing Cohorts}
\description{
This software tool is designed to generate new cohorts utilizing data from previously instantiated cohorts. It employs interval algebra operators such as UNION, INTERSECT, and MINUS to manipulate the data within the instantiated cohorts and create new cohorts.

This software tool is designed to generate new cohorts utilizing data from previously instantiated cohorts. It employs interval algebra operators such as UNION, INTERSECT, and MINUS to manipulate the data within the instantiated cohorts and create new cohorts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortAlgebra/}
  \item \url{https://CRAN.R-project.org/package=CohortAlgebra}
  \item Report bugs at \url{https://github.com/OHDSI/CohortAlgebra/issues}
}


Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortAlgebra/}
  \item \url{https://CRAN.R-project.org/package=CohortAlgebra}
  \item Report bugs at \url{https://github.com/OHDSI/CohortAlgebra/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
