% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EraFyCohortDataFrame.R
\name{eraFyCohortDataFrame}
\alias{eraFyCohortDataFrame}
\title{Era-fy cohort(s)}
\usage{
eraFyCohortDataFrame(cohort, eraconstructorpad = 0)
}
\arguments{
\item{cohort}{A data frame with cohort data conforming to cohortDefinitionId, subjectId, cohortStartDate, cohortEndDate}

\item{eraconstructorpad}{Optional value to pad cohort era construction logic. Default = 0. i.e. no padding.}
}
\value{
cohort table
}
\description{
Given a data frame in R with cohortDefinitionId, subjectId, cohortStartDate,
cohortEndDate execute era logic. Returns a cohort.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
