% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCohort.R
\name{filterCohort}
\alias{filterCohort}
\title{Filter existing cohort to create new cohort}
\usage{
filterCohort(
  connectionDetails = NULL,
  connection = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortDatabaseSchema = NULL,
  cohortTableName,
  oldCohortId,
  newCohortId,
  minCohortDays = 0,
  maxCohortDays = 9999
)
}
\arguments{
\item{connectionDetails}{(optional) Details for establishing a database connection if \code{connection} is NULL.}

\item{connection}{(optional) An existing database connection object.}

\item{tempEmulationSchema}{(optional) The schema used for emulating temporary tables;
defaults to the value set in global options with \code{getOption("sqlRenderTempEmulationSchema")}.}

\item{cohortDatabaseSchema}{The name of the schema where the cohort table exists.}

\item{cohortTableName}{The name of the cohort table.}

\item{oldCohortId}{cohort id of the cohort to filter.}

\item{newCohortId}{cohort id of the filtered cohort that will be created.}

\item{minCohortDays}{(default = 0)}

\item{maxCohortDays}{(default = 9999)}
}
\description{
This function filters an existing cohort to create a new cohort.
}
