% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateBaseCohorts.R
\name{getBaseCohortDefinitionSet}
\alias{getBaseCohortDefinitionSet}
\title{Base cohort, cohort definition set.}
\usage{
getBaseCohortDefinitionSet()
}
\description{
Base cohort, cohort definition set.
}
