% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCohortRelationship.R
\name{getCohortRelationship}
\alias{getCohortRelationship}
\title{Generate Cohort Relationship Data}
\usage{
getCohortRelationship(
  cohortTable,
  targetCohortId,
  featureCohortIds,
  closestPeriod = FALSE
)
}
\arguments{
\item{cohortTable}{A data frame with cohort data conforming to \code{cohortDefinitionId}, \code{subjectId}, \code{cohortStartDate}, \code{cohortEndDate}.}

\item{targetCohortId}{Integer value representing the cohortDefinitionId of the target cohort.}

\item{featureCohortIds}{Integer value representing the cohortDefinitionId of the feature cohort.}

\item{closestPeriod}{Boolean indicating whether to select the closest period in the feature cohort relative to the target cohort's start date. Default is FALSE.}
}
\value{
A data frame with columns \code{subjectId}, \code{startDay}, and \code{endDay}.
}
\description{
Given a data frame in R with \code{cohortDefinitionId}, \code{subjectId}, \code{cohortStartDate},
and \code{cohortEndDate}, this function processes the data to identify the relationship
between first cohort start of the target cohort, and cohort starts of the feature
cohorts. The function can optionally select the closest
period in the feature cohort relative to the target cohort's start date.
}
