% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCohortTimes.R
\name{getCohortTimes}
\alias{getCohortTimes}
\title{Create cohort summary}
\usage{
getCohortTimes(
  connectionDetails = NULL,
  connection = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cohortDatabaseSchema = NULL,
  cdmDatabaseSchema,
  cohortTable,
  cohortIds = NULL,
  cohortTableIsTemp = FALSE
)
}
\arguments{
\item{connectionDetails}{(optional) Details for establishing a database connection if \code{connection} is NULL.}

\item{connection}{(optional) An existing database connection object.}

\item{tempEmulationSchema}{(optional) The schema used for emulating temporary tables;
defaults to the value set in global options with \code{getOption("sqlRenderTempEmulationSchema")}.}

\item{cohortDatabaseSchema}{The name of the schema where the cohort table exists.}

\item{cdmDatabaseSchema}{The name of the schema with OMOP CDM person level tables.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortIds}{(optional) Vector of IDs corresponding to cohort definitions.}

\item{cohortTableIsTemp}{is Cohort table temp.}
}
\value{
A tibble with columns for cohort_definition_id, number of distinct subjects, number of events, and total days.
}
\description{
This function calculates the number of subjects, records, and total cohort days
for each provided \code{cohortDefinitionId}. It connects to a specified database,
and queries a cohort table to gather the necessary information.
}
