% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistributionOfCohortDays.R
\name{getDistributionOfCohortDays}
\alias{getDistributionOfCohortDays}
\title{Calculate cohort days summary statistics}
\usage{
getDistributionOfCohortDays(cohort, cohortDefinitionId = NULL)
}
\arguments{
\item{cohort}{A data frame that includes the columns cohortDefinitionId, subjectId, cohortStartDate, and cohortEndDate.}

\item{cohortDefinitionId}{(optional) The cohort id to filter the data; if NULL, all rows will be used.}
}
\value{
A data frame of summary statistics for each cohort definition id.
}
\description{
This function retrieves the full cohort tables into R and performs several summary statistics on the days duration of cohort memberships.
}
