% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDummyCohortDefinitionSet.R
\name{getDummyCohortDefinitionSet}
\alias{getDummyCohortDefinitionSet}
\title{Get dummy cohort definition set}
\usage{
getDummyCohortDefinitionSet(cohortId = 0)
}
\description{
Get dummy cohort definition set

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
CohortAlgebra::getDummyCohortDefinitionSet(cohortId = 0)
}

}
