% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortAlgebra-package.R, R/CohortAlgebra.R
\docType{package}
\name{CohortAlgebra-package}
\alias{CohortAlgebra}
\alias{CohortAlgebra-package}
\title{CohortAlgebra: Use of interval Algebra to create new cohort(s) from existing cohorts}
\description{
This software tool is designed to create new cohort(s) from previously instantiated cohorts. It uses interval algebra operators of UNION, Intersect and Minus on the instantiated cohorts. Cohorts are definied in the Observational Medical Outcomes Partnership (OMOP) data model format as a table with the fields cohort_definition_id, subject_id, cohort_start_date and cohort_end_date.

This software tool is designed to create new cohort(s) from previously instantiated cohorts. It uses interval algebra operators of UNION, Intersect and Minus on the instantiated cohorts. Cohorts are definied in the Observational Medical Outcomes Partnership (OMOP) data model format as a table with the fields cohort_definition_id, subject_id, cohort_start_date and cohort_end_date.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortAlgebra/}
  \item \url{https://github.com/OHDSI/CohortAlgebra}
  \item Report bugs at \url{https://github.com/OHDSI/CohortAlgebra/issues}
}


Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortAlgebra/}
  \item \url{https://github.com/OHDSI/CohortAlgebra}
  \item Report bugs at \url{https://github.com/OHDSI/CohortAlgebra/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
