% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortAlgebra.R
\docType{package}
\name{CohortAlgebra-package}
\alias{CohortAlgebra}
\alias{CohortAlgebra-package}
\title{CohortAlgebra: Cohort Algebra to create new cohort(s) from existing cohorts}
\description{
An R package that creates new cohort(s) from previously instantiated cohorts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortAlgebra/}
  \item \url{https://github.com/OHDSI/CohortAlgebra}
  \item Report bugs at \url{https://github.com/OHDSI/CohortAlgebra/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Authors:
\itemize{
  \item Adam Black \email{black@ohdsi.org}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
