% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyCohortsToTempTable.R
\name{copyCohortsToTempTable}
\alias{copyCohortsToTempTable}
\title{Copy cohorts to temp table}
\usage{
copyCohortsToTempTable(
  connection = NULL,
  oldToNewCohortId,
  sourceCohortDatabaseSchema = NULL,
  sourceCohortTable,
  targetCohortTable = "#cohort_rows",
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{oldToNewCohortId}{A data.frame object with two columns. oldCohortId and newCohortId. Both should be integers.
The oldCohortId are the cohorts that are the input cohorts that need to be transformed.
The newCohortId are the cohortIds of the corresponding output after transformation.
If the oldCohortId = newCohortId then the data corresponding to oldCohortId
will be replaced by the data from the newCohortId.}

\item{sourceCohortDatabaseSchema}{The database schema of the source cohort table.}

\item{sourceCohortTable}{The name of the source cohort table.}

\item{targetCohortTable}{A temp table to copy the cohorts from the source table.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}
}
\description{
Copy cohorts to temp table. This function is not exported.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
