% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinusCohorts.R
\name{minusCohorts}
\alias{minusCohorts}
\title{Minus cohort(s)}
\usage{
minusCohorts(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema = NULL,
  cohortTable = "cohort",
  firstCohortId,
  secondCohortId,
  newCohortId,
  purgeConflicts = FALSE,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{cohortDatabaseSchema}{Schema name where your cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cohortTable}{The name of the cohort table.}

\item{firstCohortId}{The cohort id of the cohort from which to substract.}

\item{secondCohortId}{The cohort id of the cohort that is used to substract.}

\item{newCohortId}{The cohort id of the result cohort.}

\item{purgeConflicts}{If there are conflicts in the target cohort table i.e. the target cohort table
already has records with newCohortId, do you want to purge and replace them
with transformed. By default - it will not be replaced, and an error message is thrown.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}
}
\description{
Given two cohorts, substract (minus) the dates from the first cohort, the
dates the subject also had on the second cohort.
}
\examples{
minusCohorts(
 connection = conneciton,
 cohortDatabaseSchema = "study1234",
 cohortTable = "cohort",
 firstCohortId = 1,
 secondCohortId = 2,
 newCohortId = 9,
 purgeConflicts = TRUE
)

}
